requirements.txt
flask==3.0.3
flask-login==0.6.3
flask-migrate==4.0.7
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.30
gunicorn==22.0.0
werkzeug==3.0.3
python-dotenv==1.0.1

# Azure SDKs
azure-storage-blob==12.20.0
azure-ai-formrecognizer==3.3.3
openai==1.30.1

# Database drivers
pyodbc==5.1.0
pymssql==2.3.0
